requirements.txt
beautifulsoup4
PyYAML